+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # OB-USP-AGENT Changelog since Release 8.0.0
 
+## 2024-05-13 v8.0.3
+### Added
+- USP Broker: CLI initiated gets have been optimized to pass through the path to the USP Service, when possible
+- USP Broker: Support a USP Service registering Device.DNS.SD before Device.DNS (in separate register messages)
+- USP Broker: Workaround for USP Services which have limitations on the number of parameters requested in a get
+- USP Broker: Support additional DM elements registered directly under Device.
+
+### Fixed
+- Get instances failures during path resolution should be gracefully ignored
+- An unused variable warning seen during cmake based builds has been addressed
+
+
 ## 2024-04-29 v8.0.2
 ### Added
 - VALUE_CHANGE_WILL_IGNORE flag support